requirements.txt
openai-agents
openai
python-dotenv
pydantic
requests